renderers/video/requirements.txt
Flask
Pillow
gTTS
pydub
croniter>=1.3.8  # Required for time-based scheduling
moviepy>=1.0.3
numpy>=1.24.0
